Split ReactDOM entry point
